# Conflicts:
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/common/util/ResourceProvider.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/detail/presentation/fragment/AddEditProductDetailFragment.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/detail/presentation/viewmodel/AddEditProductDetailViewModel.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/fragment/AddEditProductPreviewFragment.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModel.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/detail/presentation/viewmodel/AddEditProductDetailViewModelTest.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModelTest.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModelTestFixture.kt
#	features/merchant/product_manage_common/src/main/java/com/tokopedia/product/manage/common/feature/variant/data/model/GetProductV3.kt
#	features/merchant/product_manage_common/src/main/java/com/tokopedia/product/manage/common/feature/variant/presentation/data/GetVariantResult.kt
#	features/merchant/product_manage_common/src/main/res/layout/fragment_quick_edit_stock.xml
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/adapter/viewholder/ProductViewHolder.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/fragment/ProductManageFragment.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/mapper/ProductMapper.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/model/ViewState.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/ui/bottomsheet/ProductManageBottomSheet.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/viewmodel/ProductManageViewModel.kt
#	features/merchant/product_manage_list/src/main/res/layout/item_campaign_stock_total_editor.xml
#	features/merchant/product_manage_list/src/main/res/values/strings.xml
#	features/merchant/seller_home/src/main/java/com/tokopedia/sellerhome/view/activity/SellerHomeActivity.kt
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/constant/AccessId.kt
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/data/source/cloud/query/AuthorizeAccess.kt
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/domain/interactor/AuthorizeAccessUseCase.kt
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/domain/interactor/model/adminrevamp/AdminInfoResponse.kt
#	features/transaction/sellerorder/src/main/java/com/tokopedia/sellerorder/list/presentation/fragments/SomListFragment.kt
#	features/transaction/sellerorder/src/main/java/com/tokopedia/sellerorder/list/presentation/viewmodels/SomListViewModel.kt
#	features/transaction/sellerorder/src/test/java/com/tokopedia/sellerorder/list/presentation/viewmodels/SomListViewModelTest.kt

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/domain/usecase/AddToCartMultiLegacyUseCase.kt
@@ -44,7 +44,7 @@ class AddToCartMultiLegacyUseCase @Inject constructor(private val graphqlUseCase
                     for (jsonElement in jsonArray) {
                         try {
                             val product = jsonElement.asJsonObject
-                            val productId = product[PRODUCT_ID_KEY].asInt
+                            val productId = product[PRODUCT_ID_KEY].asLong
                             val productName = product[PRODUCT_NAME_KEY].asString
                             val quantity = product[QUANTITY_KEY].asInt
                             val productPrice = product[PRODUCT_PRICE_KEY].asString
