+++ features/user/chat_common/src/main/java/com/tokopedia/chat_common/domain/pojo/productattachment/ProductAttachmentAttributes.kt
@@ -7,7 +7,7 @@ class ProductAttachmentAttributes {
 
     @SerializedName("product_id")
     @Expose
-    val productId: Int = 0
+    val productId: Long = 0L
     @SerializedName("product_profile")
     @Expose
     val productProfile: ProductProfile = ProductProfile()
