+++ features/user/topchat/src/main/java/com/tokopedia/topchat/chatroom/view/listener/TopChatContract.kt
@@ -111,7 +111,7 @@ interface TopChatContract {
                        onError: (Throwable) -> Unit,
                        onSuccessDeleteConversation: () -> Unit)
 
-        fun getShopFollowingStatus(shopId: Int,
+        fun getShopFollowingStatus(shopId: Long,
                                    onError: (Throwable) -> Unit,
                                    onSuccessGetShopFollowingStatus: (Boolean) -> Unit)
 
@@ -158,7 +158,7 @@ interface TopChatContract {
 
         fun getStickerGroupList(chatRoom: ChatroomViewModel)
 
-        fun loadAttachmentData(msgId: Int, chatRoom: ChatroomViewModel)
+        fun loadAttachmentData(msgId: Long, chatRoom: ChatroomViewModel)
 
         fun isStickerTooltipAlreadyShow(): Boolean
